+++ GasTracker/app/src/main/java/com/twild/gastracker/Car.java
@@ -27,37 +27,37 @@ public class Car
         tripList = new ArrayList<>();
     }
 
-    public void addFillup(int day, int month, int year, int mileage, double amount, double price, int full)
+    public void addFillup(int day, int month, int year, double mileage, double amount, double price, int full)
     {
         fillUpList.add(new Fillup(day, month, year, mileage, amount, price, full));
         Collections.sort(fillUpList);
     }
 
-    public void editFillup(int index, int day, int month, int year, int mileage, double amount, double price, int full)
+    public void editFillup(int index, int day, int month, int year, double mileage, double amount, double price, int full)
     {
         fillUpList.set(index, new Fillup(day, month, year, mileage, amount, price, full));
         Collections.sort(fillUpList);
     }
 
-    public void addMaintenance(int day, int month, int year, int mileage, String type, String info)
+    public void addMaintenance(int day, int month, int year, double mileage, String type, String info)
     {
         maintenanceList.add(new Maintenance(day, month, year, mileage, type, info));
         Collections.sort(maintenanceList);
     }
 
-    public void editMaintenance(int index, int day, int month, int year, int mileage, String type, String info)
+    public void editMaintenance(int index, int day, int month, int year, double mileage, String type, String info)
     {
         maintenanceList.set(index, new Maintenance(day, month, year, mileage, type, info));
         Collections.sort(maintenanceList);
     }
 
-    public void addTrip(int day, int month, int year, int mileage, int endMileage, String info)
+    public void addTrip(int day, int month, int year, double mileage, double endMileage, String info)
     {
         tripList.add(new Trip(day, month, year, mileage, endMileage, info));
         Collections.sort(tripList);
     }
 
-    public void editTrip(int index, int day, int month, int year, int mileage, int endMileage, String info)
+    public void editTrip(int index, int day, int month, int year, double mileage, double endMileage, String info)
     {
         tripList.set(index, new Trip(day, month, year, mileage, endMileage, info));
         Collections.sort(tripList);
